set the pid parameter / Project intermit for a short time
Will the micro-quadrocoper need control to fly straight up or down? The
pid and the romote control add to the double-loop test, still need to
adjust the parameters to fly stable and safely.
some problems:
The 720 motor broken down easily.
The battery's weight need comfirm.
The control optition is quite significant.